chore(electrum,testenv): move `test_reorg_is_detected_in_electrsd`
